+++ lib/src/timeline/timeline_item.dart
@@ -5,6 +5,7 @@ class TimelineItem extends StatelessWidget {
   const TimelineItem(
       {super.key,
       required this.colors,
+      required this.lang,
       required this.index,
       required this.centerItemIndex,
       required this.nowIndex,
@@ -16,6 +17,7 @@ class TimelineItem extends StatelessWidget {
       required this.openDayDetail});
 
   final Map<String, Color> colors;
+  final String lang;
   final int index;
   final int centerItemIndex;
   final int nowIndex;
@@ -31,8 +33,8 @@ class TimelineItem extends StatelessWidget {
     DateTime date = days[index]['date'];
 
     double dayProgress = 0;
-    if (days[index] != null && days[index]['workLoadLevel'] != null && days[index]['workLoadLevel'] > 0) {
-      dayProgress = 100 * days[index]['completedLevel'] / days[index]['workLoadLevel'];
+    if (days[index] != null && days[index]['busyEffort'] != null && days[index]['busyEffort'] > 0) {
+      dayProgress = 100 * days[index]['completedEffort'] / days[index]['busyEffort'];
     }
 
     return Align(
@@ -48,7 +50,7 @@ class TimelineItem extends StatelessWidget {
             child: SizedBox(
                 width: dayWidth - dayMargin,
                 height: (days[index]['capacityLevelMax'] > 0
-                        ? ((height * days[index]['capacityLevel']) /
+                        ? ((height * days[index]['capacityEffort']) /
                             days[index]['capacityLevelMax'])
                         : height) +
                     50,
@@ -61,7 +63,7 @@ class TimelineItem extends StatelessWidget {
                             bottom: dayMargin / 3),
                         width: dayWidth - dayMargin,
                         height: (days[index]['capacityLevelMax'] > 0
-                                ? ((height * days[index]['capacityLevel']) /
+                                ? ((height * days[index]['capacityEffort']) /
                                     days[index]['capacityLevelMax'])
                                 : height) - 20,
                         decoration: BoxDecoration(
@@ -78,11 +80,11 @@ class TimelineItem extends StatelessWidget {
                                   alignment: Alignment.bottomCenter,
                                   child: Container(
                                       width: dayWidth - dayMargin,
-                                      height: days[index]['workLoadLevel'] > 0
+                                      height: days[index]['busyEffort'] > 0
                                           ? ((height *
                                                   days[index]
-                                                      ['completedLevel']) /
-                                              days[index]['workLoadLevel'])
+                                                      ['completedEffort']) /
+                                              days[index]['busyEffort'])
                                           : height,
                                       decoration: BoxDecoration(
                                         borderRadius: BorderRadius.circular(10),
@@ -90,33 +92,8 @@ class TimelineItem extends StatelessWidget {
                                             ? colors['accent1']
                                             : colors['primary'],
                                       ),
-                                      child: Column(
-                                        mainAxisAlignment:
-                                            MainAxisAlignment.end,
-                                        children: [
-                                          if (isMultiproject)
-                                            Align(
-                                                alignment:
-                                                    Alignment.bottomCenter,
-                                                child: Container(
-                                                    width: dayWidth - dayMargin,
-                                                    height: days[index][
-                                                                'workLoadLevel'] >
-                                                            0
-                                                        ? ((height *
-                                                                days[index][
-                                                                    'myCompletedLevel']) /
-                                                            days[index][
-                                                                'workLoadLevel'])
-                                                        : height,
-                                                    decoration: BoxDecoration(
-                                                      borderRadius:
-                                                          BorderRadius.circular(
-                                                              10),
-                                                      color: colors['primary'],
-                                                    )))
-                                        ],
-                                      )))
+                                  )
+                              )
                             else
                               Center(
                                   child: Icon(Icons.sunny,
@@ -124,7 +101,7 @@ class TimelineItem extends StatelessWidget {
                           ],
                         )),
                     Text(
-                      DateFormat.E('fr_FR').format(date),
+                      DateFormat.E(lang).format(date),
                       style: TextStyle(
                           color: colors['primaryText'],
                           fontWeight: centerItemIndex == index